+++ MelonLoader/Extensions/PatchShield.cs
@@ -24,29 +24,28 @@ namespace MelonLoader
 			Type patchProcessorType = typeof(PatchProcessor);
 			Type patchShieldType = typeof(PatchShield);
 			PatchProcessor_OriginalRef = AccessTools.FieldRefAccess<MethodBase>(patchProcessorType, "original");
-			HarmonyLib.Harmony harmonyInstance = new HarmonyLib.Harmony("PatchShield");
 
 			try
 			{
-				harmonyInstance.Patch(
+				Core.HarmonyInstance.Patch(
 					AccessTools.Method("HarmonyLib.PatchFunctions:ReversePatch"),
-					new HarmonyMethod(AccessTools.Method(patchShieldType, "PatchMethod_PatchFunctions_ReversePatch"))
+					AccessTools.Method(patchShieldType, "PatchMethod_PatchFunctions_ReversePatch").ToNewHarmonyMethod()
 					);
 			}
 			catch (Exception ex) { LogException(ex); }
 
 			try
 			{
-				HarmonyMethod unpatchMethod = new HarmonyMethod(AccessTools.Method(patchShieldType, "PatchMethod_PatchProcessor_Unpatch"));
+				HarmonyMethod unpatchMethod = AccessTools.Method(patchShieldType, "PatchMethod_PatchProcessor_Unpatch").ToNewHarmonyMethod();
 				foreach (MethodInfo method in patchProcessorType.GetMethods(BindingFlags.Public | BindingFlags.Instance).Where(x => x.Name.Equals("Unpatch")))
-					harmonyInstance.Patch(method, unpatchMethod);
+					Core.HarmonyInstance.Patch(method, unpatchMethod);
 			}
 			catch (Exception ex) { LogException(ex); }
 
 			try
 			{
-				harmonyInstance.Patch(AccessTools.Method(patchProcessorType, "Patch"),
-					new HarmonyMethod(AccessTools.Method(patchShieldType, "PatchMethod_PatchProcessor_Patch"))
+				Core.HarmonyInstance.Patch(AccessTools.Method(patchProcessorType, "Patch"),
+					AccessTools.Method(patchShieldType, "PatchMethod_PatchProcessor_Patch").ToNewHarmonyMethod()
 					);
 			}
 			catch (Exception ex) { LogException(ex); }