+++ MelonLoader/Melons/Attributes/Priority.cs
@@ -13,6 +13,8 @@ namespace MelonLoader
         public MelonPriorityAttribute(int priority = 0) => Priority = priority;
     }
 
+    /*
     [AttributeUsage(AttributeTargets.Assembly)]
     public class MelonPriority : MelonPriorityAttribute { public MelonPriority(int priority = 0) : base(priority) { } }
+    */
 }